move categories to separate folders
refactor!: move categories to en and tr folders

feat!: added in daisy ui plugin via tailwind.config.ts
initial commit testing for fetching mongodb data via server actions
