+++ src/actions/news.actions.ts
@@ -1,14 +1,14 @@
-'use server'
+"use server";
 
 import { connect } from "@/lib/db";
 import News from "@/models/news.model";
 
 export async function getNews() {
-    try {
-        await connect();
-        const news = await News.find();
-        return news;
-    } catch (error) {
-        throw new Error("Failed to fetch news");
-    }
-} 
+  try {
+    await connect();
+    const news = await News.find();
+    return news;
+  } catch (error) {
+    throw new Error("Failed to fetch news");
+  }
+}
