+++ src/app/(categories)/(en)/discounts/page.tsx
