+++ src/app/(categories)/(en)/entertainment/page.tsx
