+++ src/app/(categories)/(en)/scholarships/page.tsx
