+++ src/app/(categories)/(en)/stores/page.tsx
