+++ src/app/(categories)/(tr)/etkinlikler/page.tsx
