+++ src/app/(categories)/(tr)/kafeler/page.tsx
