+++ src/app/(categories)/(tr)/yurtlar/page.tsx
