+++ src/components/ui/CategoryCard/CategoryBody/CategoryBody.tsx
@@ -1,0 +1,13 @@
+import { ReactNode } from "react";
+
+interface CardBody {
+  children?: ReactNode;
+}
+
+export default function CategoryBody({ children }: CardBody) {
+  return (
+    <>
+      <div className="space-y-8">{children}</div>
+    </>
+  );
+}
