+++ src/components/ui/CategoryCard/CategoryHeader/CategoryHeader.tsx
@@ -1,0 +1,3 @@
+export default function CategoryHeader({ title }: { title: string }) {
+  return <h1 className="text-4xl font-bold text-gray-900 mb-6">{title}</h1>;
+}
